# docs/protocol/openapi.v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Copilot Bridge Cloud Desktop API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://ai.mddxz.top"
    },
    {
      "url": "http://127.0.0.1:3001"
    }
  ],
  "components": {
    "securitySchemes": {
      "DesktopBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "summary": "Desktop login and device activation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string"
                  },
                  "device": {
                    "type": "object",
                    "properties": {
                      "deviceId": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "deviceName": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 120
                      },
                      "platform": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 60
                      },
                      "osVersion": {
                        "type": "string",
                        "maxLength": 120,
                        "default": ""
                      },
                      "appVersion": {
                        "type": "string",
                        "maxLength": 120,
                        "default": ""
                      }
                    },
                    "required": [
                      "deviceId",
                      "deviceName",
                      "platform"
                    ]
                  }
                },
                "required": [
                  "email",
                  "password",
                  "device"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accessToken": {
                      "type": "string"
                    },
                    "refreshToken": {
                      "type": "string"
                    },
                    "expiresIn": {
                      "type": "number",
                      "enum": [
                        1800
                      ]
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "email": {
                          "type": "string",
                          "format": "email"
                        },
                        "role": {
                          "type": "string",
                          "enum": [
                            "USER",
                            "ADMIN"
                          ]
                        },
                        "status": {
                          "type": "string",
                          "enum": [
                            "ACTIVE",
                            "DISABLED",
                            "EXPIRED"
                          ]
                        }
                      },
                      "required": [
                        "id",
                        "email",
                        "role",
                        "status"
                      ]
                    },
                    "device": {
                      "type": "object",
                      "properties": {
                        "deviceId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "deviceName": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 120
                        },
                        "platform": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 60
                        },
                        "osVersion": {
                          "type": "string",
                          "maxLength": 120,
                          "default": ""
                        },
                        "appVersion": {
                          "type": "string",
                          "maxLength": 120,
                          "default": ""
                        },
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "userId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "status": {
                          "type": "string",
                          "enum": [
                            "ACTIVE",
                            "REVOKED"
                          ]
                        },
                        "activatedAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "lastSeenAt": {
                          "type": [
                            "string",
                            "null"
                          ],
                          "format": "date-time"
                        }
                      },
                      "required": [
                        "deviceId",
                        "deviceName",
                        "platform",
                        "id",
                        "userId",
                        "status",
                        "activatedAt",
                        "lastSeenAt"
                      ]
                    }
                  },
                  "required": [
                    "accessToken",
                    "refreshToken",
                    "expiresIn",
                    "user",
                    "device"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "UNAUTHORIZED",
                            "TOKEN_EXPIRED",
                            "DEVICE_NOT_REGISTERED",
                            "DEVICE_REVOKED",
                            "DEVICE_LIMIT_REACHED",
                            "ACCOUNT_DISABLED",
                            "SUBSCRIPTION_REQUIRED",
                            "SUBSCRIPTION_EXPIRED",
                            "MODEL_NOT_ALLOWED",
                            "MONTHLY_QUOTA_EXCEEDED",
                            "RATE_LIMITED",
                            "PROVIDER_UNAVAILABLE",
                            "MODEL_UNAVAILABLE",
                            "GATEWAY_TIMEOUT",
                            "CLIENT_THREAD_ID_REQUIRED",
                            "VALIDATION_ERROR",
                            "INTERNAL_ERROR"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ]
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "UNAUTHORIZED",
                            "TOKEN_EXPIRED",
                            "DEVICE_NOT_REGISTERED",
                            "DEVICE_REVOKED",
                            "DEVICE_LIMIT_REACHED",
                            "ACCOUNT_DISABLED",
                            "SUBSCRIPTION_REQUIRED",
                            "SUBSCRIPTION_EXPIRED",
                            "MODEL_NOT_ALLOWED",
                            "MONTHLY_QUOTA_EXCEEDED",
                            "RATE_LIMITED",
                            "PROVIDER_UNAVAILABLE",
                            "MODEL_UNAVAILABLE",
                            "GATEWAY_TIMEOUT",
                            "CLIENT_THREAD_ID_REQUIRED",
                            "VALIDATION_ERROR",
                            "INTERNAL_ERROR"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ]
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Access denied",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "UNAUTHORIZED",
                            "TOKEN_EXPIRED",
                            "DEVICE_NOT_REGISTERED",
                            "DEVICE_REVOKED",
                            "DEVICE_LIMIT_REACHED",
                            "ACCOUNT_DISABLED",
                            "SUBSCRIPTION_REQUIRED",
                            "SUBSCRIPTION_EXPIRED",
                            "MODEL_NOT_ALLOWED",
                            "MONTHLY_QUOTA_EXCEEDED",
                            "RATE_LIMITED",
                            "PROVIDER_UNAVAILABLE",
                            "MODEL_UNAVAILABLE",
                            "GATEWAY_TIMEOUT",
                            "CLIENT_THREAD_ID_REQUIRED",
                            "VALIDATION_ERROR",
                            "INTERNAL_ERROR"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ]
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "429": {
            "description": "Quota or rate limit",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "UNAUTHORIZED",
                            "TOKEN_EXPIRED",
                            "DEVICE_NOT_REGISTERED",
                            "DEVICE_REVOKED",
                            "DEVICE_LIMIT_REACHED",
                            "ACCOUNT_DISABLED",
                            "SUBSCRIPTION_REQUIRED",
                            "SUBSCRIPTION_EXPIRED",
                            "MODEL_NOT_ALLOWED",
                            "MONTHLY_QUOTA_EXCEEDED",
                            "RATE_LIMITED",
                            "PROVIDER_UNAVAILABLE",
                            "MODEL_UNAVAILABLE",
                            "GATEWAY_TIMEOUT",
                            "CLIENT_THREAD_ID_REQUIRED",
                            "VALIDATION_ERROR",
                            "INTERNAL_ERROR"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ]
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}